fix http request parsing issues
Extensions like [cph-submit](https://github.com/agrawal-d/cph-submit)
will periodically make GET requests to the HTTP server to poll for
code submission. If we want to support that protocol, then we must
have a mechanism to differentiate between the two types of requests.
Currently, we are only checking if `method == "POST"`, but since the
HTTP request is completely parsed, additional stricter conditions can
be imposed.

The HTTP parser is AI generated, but it works fine on my machine and
the code is readable.