BREAKING CHANGE: code transform callback signature
The new transform callback now takes two parameter: the original code
and a submit callback. Instead of returning the transformed code, the
submit callback should be called with the transformed code as its
parameter.

In the old API, the transform callback is more restricted, since
it could not use some asynchronous API, so a callback as simple as
piping the old code to an external code transformer is unsupported.